move worker to seperate csproj
