+++ starsky/starsky.foundation.sync/SyncInterfaces/IManualBackgroundSyncService.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+using starsky.foundation.database.Models;
+
+namespace starsky.foundation.sync.SyncInterfaces
+{
+	public interface IManualBackgroundSyncService
+	{
+		Task<FileIndexItem.ExifStatus> ManualSync(string subPath);
+	}
+}
