+++ starsky/starsky.foundation.worker/starsky.foundation.worker.csproj
@@ -1,0 +1,21 @@
+<Project Sdk="Microsoft.NET.Sdk">
+    
+    <PropertyGroup>
+        <TargetFramework>netstandard2.0</TargetFramework>
+        <!-- SonarQube needs this -->
+        <ProjectGuid>{240fbcdb-4379-4bbe-b154-eb4f87504d9e}</ProjectGuid>
+        <DebugType>Full</DebugType>
+        <Version>0.4.8</Version>
+        <RootNamespace>starsky.foundation.worker</RootNamespace>
+    </PropertyGroup>
+    
+    <ItemGroup>
+      <ProjectReference Include="..\starsky.foundation.injection\starsky.foundation.injection.csproj" />
+      <ProjectReference Include="..\starsky.foundation.webtelemetry\starsky.foundation.webtelemetry.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="3.1.13" />
+    </ItemGroup>
+
+</Project>
