+++ starsky/starsky/Controllers/ApiController.cs
@@ -4,8 +4,8 @@ using Microsoft.Extensions.Caching.Memory;
 using starsky.foundation.database.Interfaces;
 using starsky.foundation.platform.Models;
 using starsky.foundation.storage.Interfaces;
+using starsky.foundation.worker.Services;
 using starsky.foundation.writemeta.Interfaces;
-using starskycore.Services;
 
 namespace starsky.Controllers
 {
