+++ starsky/starsky/Controllers/GeoController.cs
@@ -8,7 +8,6 @@ using starsky.feature.geolookup.Interfaces;
 using starsky.feature.geolookup.Models;
 using starsky.feature.geolookup.Services;
 using starsky.foundation.database.Models;
-using starskycore.Services;
 using starsky.foundation.platform.Helpers;
 using starsky.foundation.platform.Models;
 using starsky.foundation.readmeta.Interfaces;
@@ -17,8 +16,8 @@ using starsky.foundation.storage.Interfaces;
 using starsky.foundation.storage.Models;
 using starsky.foundation.storage.Services;
 using starsky.foundation.storage.Storage;
+using starsky.foundation.worker.Services;
 using starsky.foundation.writemeta.Interfaces;
-using starsky.foundation.writemeta.Services;
 
 namespace starsky.Controllers
 {
