+++ starsky/starsky/Controllers/PublishController.cs
@@ -11,7 +11,7 @@ using starsky.foundation.platform.Helpers;
 using starsky.foundation.platform.Models;
 using starsky.foundation.storage.Interfaces;
 using starsky.foundation.storage.Storage;
-using starskycore.Services;
+using starsky.foundation.worker.Services;
 
 namespace starsky.Controllers
 {
