+++ starsky/starsky/Controllers/SynchronizeController.cs
@@ -1,23 +1,27 @@
+using System;
 using System.Collections.Generic;
+using System.Text.Json;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Caching.Memory;
 using starsky.foundation.database.Interfaces;
 using starsky.foundation.database.Models;
+using starsky.foundation.platform.JsonConverter;
+using starsky.foundation.realtime.Interfaces;
 using starsky.foundation.sync.SyncInterfaces;
+using starsky.foundation.worker.Services;
 
 namespace starsky.Controllers
 {
 	[Authorize]
 	public class SynchronizeController : Controller
 	{
-		private readonly ISynchronize _synchronize;
-		private readonly IQuery _query;
+		private readonly IManualBackgroundSyncService _manualBackgroundSyncService;
 
-		public SynchronizeController(ISynchronize synchronize, IQuery query )
+		public SynchronizeController(IManualBackgroundSyncService manualBackgroundSyncService)
 		{
-			_synchronize = synchronize;
-			_query = query;
+			_manualBackgroundSyncService = manualBackgroundSyncService;
 		}
 
 		/// <summary>
@@ -34,18 +38,15 @@ namespace starsky.Controllers
 		[Produces("application/json")]	   
 		public async Task<IActionResult> Index(string f)
 		{
-			var fileIndexItem = await _query.GetObjectByFilePathAsync(f);
-			if ( fileIndexItem == null )
+			var status = await _manualBackgroundSyncService.ManualSync(f);
+			switch ( status )
 			{
-				return NotFound(new List<FileIndexItem>
-				{
-					new FileIndexItem(f)
-					{
-						Status = FileIndexItem.ExifStatus.NotFoundNotInIndex
-					}
-				});
+				case FileIndexItem.ExifStatus.NotFoundNotInIndex:
+					return NotFound("Failed");
+				case FileIndexItem.ExifStatus.OperationNotSupported:
+					return BadRequest("Already started");
 			}
-			return Ok(await _synchronize.Sync(fileIndexItem.FilePath));
+			return Ok("Job created");
 		}
 	}
 }
