+++ starsky/starskytest/Controllers/GeoControllerTest.cs
@@ -15,11 +15,10 @@ using starsky.foundation.database.Interfaces;
 using starsky.foundation.database.Models;
 using starsky.foundation.database.Query;
 using starsky.foundation.platform.Extensions;
-using starskycore.Services;
 using starsky.foundation.platform.Models;
 using starsky.foundation.readmeta.Interfaces;
+using starsky.foundation.worker.Services;
 using starsky.foundation.writemeta.Interfaces;
-using starsky.foundation.writemeta.Services;
 using starskytest.FakeCreateAn;
 using starskytest.FakeMocks;
 using starskytest.Models;
