+++ starsky/starskytest/Controllers/MetaUpdateControllerTest.cs
@@ -23,8 +23,8 @@ using starsky.foundation.readmeta.Interfaces;
 using starsky.foundation.storage.Interfaces;
 using starsky.foundation.storage.Services;
 using starsky.foundation.storage.Storage;
+using starsky.foundation.worker.Services;
 using starsky.foundation.writemeta.Interfaces;
-using starskycore.Services;
 using starskytest.FakeCreateAn;
 using starskytest.FakeMocks;
 using starskytest.Models;
