+++ starsky/starskytest/Controllers/SyncControllerTest.cs
@@ -20,9 +20,9 @@ using starsky.foundation.platform.Models;
 using starsky.foundation.readmeta.Interfaces;
 using starsky.foundation.storage.Interfaces;
 using starsky.foundation.storage.Services;
+using starsky.foundation.worker.Services;
 using starsky.foundation.writemeta.Interfaces;
 using starskycore.Interfaces;
-using starskycore.Services;
 using starskycore.ViewModels;
 using starskytest.FakeCreateAn;
 using starskytest.FakeMocks;
